refactor: also extract tooltip overlay logic to mixin
