+++ packages/tooltip/src/vaadin-tooltip-overlay.js
@@ -6,10 +6,9 @@
 import { html, PolymerElement } from '@polymer/polymer/polymer-element.js';
 import { defineCustomElement } from '@vaadin/component-base/src/define.js';
 import { DirMixin } from '@vaadin/component-base/src/dir-mixin.js';
-import { OverlayMixin } from '@vaadin/overlay/src/vaadin-overlay-mixin.js';
-import { PositionMixin } from '@vaadin/overlay/src/vaadin-overlay-position-mixin.js';
 import { overlayStyles } from '@vaadin/overlay/src/vaadin-overlay-styles.js';
 import { registerStyles, ThemableMixin } from '@vaadin/vaadin-themable-mixin/vaadin-themable-mixin.js';
+import { TooltipOverlayMixin } from './vaadin-tooltip-overlay-mixin.js';
 import { tooltipOverlayStyles } from './vaadin-tooltip-overlay-styles.js';
 
 registerStyles('vaadin-tooltip-overlay', [overlayStyles, tooltipOverlayStyles], {
@@ -22,12 +21,11 @@ registerStyles('vaadin-tooltip-overlay', [overlayStyles, tooltipOverlayStyles], 
  * @customElement
  * @extends HTMLElement
  * @mixes DirMixin
- * @mixes OverlayMixin
- * @mixes PositionMixin
  * @mixes ThemableMixin
+ * @mixes TooltipOverlayMixin
  * @private
  */
-class TooltipOverlay extends PositionMixin(OverlayMixin(DirMixin(ThemableMixin(PolymerElement)))) {
+class TooltipOverlay extends TooltipOverlayMixin(DirMixin(ThemableMixin(PolymerElement))) {
   static get is() {
     return 'vaadin-tooltip-overlay';
   }
@@ -41,15 +39,6 @@ class TooltipOverlay extends PositionMixin(OverlayMixin(DirMixin(ThemableMixin(P
     `;
   }
 
-  static get properties() {
-    return {
-      position: {
-        type: String,
-        reflectToAttribute: true,
-      },
-    };
-  }
-
   /** @protected */
   ready() {
     super.ready();
@@ -59,66 +48,6 @@ class TooltipOverlay extends PositionMixin(OverlayMixin(DirMixin(ThemableMixin(P
     // so querySelector() would return null. So we use this workaround to set properties.
     this.owner = this.__dataHost;
     this.owner._overlayElement = this;
-  }
-
-  requestContentUpdate() {
-    super.requestContentUpdate();
-
-    this.toggleAttribute('hidden', this.textContent.trim() === '');
-
-    // Copy custom properties from the tooltip
-    if (this.positionTarget && this.owner) {
-      const style = getComputedStyle(this.owner);
-      ['top', 'bottom', 'start', 'end'].forEach((prop) => {
-        this.style.setProperty(
-          `--vaadin-tooltip-offset-${prop}`,
-          style.getPropertyValue(`--vaadin-tooltip-offset-${prop}`),
-        );
-      });
-    }
-  }
-
-  /**
-   * @protected
-   * @override
-   */
-  _updatePosition() {
-    super._updatePosition();
-
-    if (!this.positionTarget) {
-      return;
-    }
-
-    // Center the tooltip overlay horizontally
-    if (this.position === 'bottom' || this.position === 'top') {
-      const targetRect = this.positionTarget.getBoundingClientRect();
-      const overlayRect = this.$.overlay.getBoundingClientRect();
-
-      const offset = targetRect.width / 2 - overlayRect.width / 2;
-
-      if (this.style.left) {
-        const left = overlayRect.left + offset;
-        if (left > 0) {
-          this.style.left = `${left}px`;
-        }
-      }
-
-      if (this.style.right) {
-        const right = parseFloat(this.style.right) + offset;
-        if (right > 0) {
-          this.style.right = `${right}px`;
-        }
-      }
-    }
-
-    // Center the tooltip overlay vertically
-    if (this.position === 'start' || this.position === 'end') {
-      const targetRect = this.positionTarget.getBoundingClientRect();
-      const overlayRect = this.$.overlay.getBoundingClientRect();
-
-      const offset = targetRect.height / 2 - overlayRect.height / 2;
-      this.style.top = `${overlayRect.top + offset}px`;
-    }
   }
 }
